Trekker ut Tilpasning av skjemaer til endringer andre steder til egne klasser
