+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/steg/VilkårsvurderingSteg.kt
@@ -8,7 +8,7 @@ import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ba.sak.kjerne.beregning.BeregningService
 import no.nav.familie.ba.sak.kjerne.eøs.felles.BehandlingId
-import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.KompetanseService
+import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.TilpassKompetanserTilRegelverkService
 import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.VilkårService
 import no.nav.familie.ba.sak.kjerne.vilkårsvurdering.validerIkkeBlandetRegelverk
@@ -24,7 +24,7 @@ class VilkårsvurderingSteg(
     private val beregningService: BeregningService,
     private val persongrunnlagService: PersongrunnlagService,
     private val tilbakestillBehandlingService: TilbakestillBehandlingService,
-    private val kompetanseService: KompetanseService,
+    private val tilpassKompetanserTilRegelverkService: TilpassKompetanserTilRegelverkService,
     private val featureToggleService: FeatureToggleService
 ) : BehandlingSteg<String> {
 
@@ -70,7 +70,7 @@ class VilkårsvurderingSteg(
         beregningService.oppdaterBehandlingMedBeregning(behandling, personopplysningGrunnlag)
 
         if (featureToggleService.isEnabled(FeatureToggleConfig.KAN_BEHANDLE_EØS)) {
-            kompetanseService.tilpassKompetanserTilRegelverk(BehandlingId(behandling.id))
+            tilpassKompetanserTilRegelverkService.tilpassKompetanserTilRegelverk(BehandlingId(behandling.id))
         }
 
         behandlingstemaService.oppdaterBehandlingstema(
